move single_dispatch to its own file
