+++ src/ssort/_method_requirements.py
@@ -4,7 +4,7 @@ import ast
 from typing import Iterable
 
 from ssort._ast import iter_child_nodes
-from ssort._utils import single_dispatch
+from ssort._single_dispatch import single_dispatch
 
 
 @single_dispatch
