+++ src/ssort/_single_dispatch.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+import functools
+from typing import Any, Callable, Generic, TypeVar
+
+_T = TypeVar("_T")
+
+
+__all__ = ["single_dispatch"]
+
+
+class _SingleDispatch(Generic[_T]):
+    """A more performant implementation of functools.singledispatch."""
+
+    def __init__(self, function: Callable[..., _T]) -> None:
+        functools.update_wrapper(self, function)
+        self._function: Callable[..., _T] = function
+        self._functions: dict[type[Any], Callable[..., _T]] = {}
+
+    def register(
+        self, cls: type[Any]
+    ) -> Callable[[Callable[..., _T]], Callable[..., _T]]:
+        def decorator(function: Callable[..., _T]) -> Callable[..., _T]:
+            self._functions[cls] = function
+            return function
+
+        return decorator
+
+    def __call__(self, arg: Any, *args: Any) -> _T:
+        return self._functions.get(type(arg), self._function)(arg, *args)
+
+
+single_dispatch = _SingleDispatch
